The Menu class was created to represent a menu interface. Game1 has been changed so that the basic setup required to Draw the three main menus is implemented.
STILL NEEDED:
A placeholder image is required for each individual menu.
No characters or MapObjects are drawn, as more information is required in regards to keeping track of the locations of each object, and what Characters / MapObjects are located within each room.
OR, an easier fix may be a way of tracking which room is the current active room so that room can me drawn along with the contents within it.